+++ user-api/src/main/java/com/zerobase/commerceproject/user/UserApplication.java
@@ -1,4 +1,4 @@
-package com.zerobase.commerce_project.user;
+package com.zerobase.commerceproject.user;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -11,7 +11,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @EnableJpaRepositories
 @SpringBootApplication
 public class UserApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(UserApplication.class, args);
-    }
+
+  public static void main(String[] args) {
+    SpringApplication.run(UserApplication.class, args);
+  }
 }
