+++ user-api/src/main/java/com/zerobase/commerceproject/user/client/MailgunClient.java
@@ -1,6 +1,6 @@
-package com.zerobase.commerce_project.user.client;
+package com.zerobase.commerceproject.user.client;
 
-import com.zerobase.commerce_project.user.client.mailgun.SendMailForm;
+import com.zerobase.commerceproject.user.client.mailgun.SendMailForm;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.SpringQueryMap;
