+++ user-api/src/main/java/com/zerobase/commerceproject/user/config/FeignConfig.java
@@ -1,4 +1,4 @@
-package com.zerobase.commerce_project.user.config;
+package com.zerobase.commerceproject.user.config;
 
 import feign.auth.BasicAuthRequestInterceptor;
 import org.springframework.beans.factory.annotation.Qualifier;
